+++ src/grid_doctor/swath/__init__.py
@@ -1,4 +1,4 @@
-"""Point-binning helpers for swath and other point-sampled data.
+"""Point-binning for swath and other point-sampled data.
 
 This module maps *point-sampled* data — satellite Level-2 swaths, station
 records, trajectories, or any dataset where every sample carries its own
@@ -36,296 +36,53 @@ The output of the dense path carries the full grid-doctor metadata
 [`coarsen_healpix`][grid_doctor.helpers.coarsen_healpix],
 [`save_pyramid`][grid_doctor.helpers.save_pyramid], and CF-aware viewers
 work unchanged.
+
+Acknowledgments
+---------------
+The binning approach — direct cell assignment via ``healpix_geo`` plus
+grouped reduction with ``numpy_groupies``, including mode aggregation
+through per-class counting — was informed by earlier EarthCARE-to-HEALPix
+conversion work by Anne Fouilloux
+(https://github.com/annefou/earthcare-dggs, MIT licensed).  This module
+is an independent implementation and deliberately diverges on cell
+indexing (perfect sphere instead of WGS84, for hub-wide alignment) and
+fill-value handling (declared attributes instead of dtype conventions).
+
+The initial implementation of this module, its test suite, and the
+accompanying documentation were largely generated by Claude, an AI
+assistant made by Anthropic, working from the grid-doctor codebase and
+the Waterpark technical-decisions document; the code was reviewed,
+adapted, and is maintained by the grid-doctor developers.
 """
 
 from __future__ import annotations
 
 import logging
 from collections.abc import Mapping
-from typing import cast
 
 import numpy as np
-import numpy_groupies as npg
 import xarray as xr
 
-from .remap import _attach_healpix_coords
-from .remap_backend import (
-    _LAT_NAMES,
-    _LON_NAMES,
+from ..remap import _attach_healpix_coords
+from ..remap_backend import (
     _canonical_lon,
-    _normalise_angle_units,
     _require_healpix_geo_module,
 )
-from .types import BinAgg, FloatArray, Int64Array, SourceUnits
+from ..types import BinAgg, Int64Array, SourceUnits
+from .utils import (
+    AGG_TO_METHOD,
+    FILL_ATTR_NAMES,
+    MAX_NESTED_LEVEL,
+    NPG_FUNC,
+    bin_count,
+    bin_mode,
+    bin_simple,
+    masked_float64,
+    resolve_methods,
+    resolve_point_coords,
+)
 
 logger = logging.getLogger(__name__)
-
-_MAX_NESTED_LEVEL = 29
-"""Maximum depth supported by 64-bit nested HEALPix indices."""
-
-_AGG_TO_METHOD: dict[str, str] = {
-    "mean": "binned-mean",
-    "mode": "binned-mode",
-    "min": "binned-min",
-    "max": "binned-max",
-    "count": "binned-count",
-}
-"""Mapping from aggregation name to the ``grid_doctor_method`` attribute."""
-
-_NPG_FUNC: dict[str, str] = {
-    "mean": "nanmean",
-    "min": "nanmin",
-    "max": "nanmax",
-}
-"""NaN-aware numpy-groupies reductions for the simple aggregations."""
-
-_FILL_ATTR_NAMES: tuple[str, ...] = ("_FillValue", "missing_value")
-"""Variable attributes inspected for fill values, in priority order."""
-
-
-# ===================================================================
-# Coordinate resolution
-# ===================================================================
-
-
-def _find_coord_name(
-    ds: xr.Dataset,
-    candidates: tuple[str, ...],
-    explicit: str | None,
-    kind: str,
-) -> str:
-    """Return the name of a coordinate variable in *ds*.
-
-    Args:
-        ds: Source dataset.
-        candidates: Priority-ordered candidate names.
-        explicit: Explicit name that overrides auto-detection.
-        kind: ``"latitude"`` or ``"longitude"`` (for error messages).
-
-    Returns:
-        The resolved variable name.
-
-    Raises:
-        KeyError: When an explicit name is not present in *ds*.
-        ValueError: When no candidate name is found.
-    """
-    if explicit is not None:
-        if explicit not in ds.coords and explicit not in ds.data_vars:
-            raise KeyError(f"{kind} variable {explicit!r} not found in dataset.")
-        return explicit
-    for name in candidates:
-        if name in ds.coords or name in ds.data_vars:
-            return name
-    available = sorted({*map(str, ds.coords), *map(str, ds.data_vars)})
-    raise ValueError(
-        f"Could not locate a {kind} coordinate. Pass "
-        f"``{kind[:3]}_name=...`` explicitly. Available names: {available}."
-    )
-
-
-def _resolve_point_coords(
-    ds: xr.Dataset,
-    *,
-    lat_name: str | None,
-    lon_name: str | None,
-    source_units: SourceUnits,
-) -> tuple[FloatArray, FloatArray, tuple[str, ...]]:
-    """Extract per-sample coordinates and the sample dimensions.
-
-    Args:
-        ds: Source dataset.
-        lat_name: Explicit latitude variable name or *None*.
-        lon_name: Explicit longitude variable name or *None*.
-        source_units: Angular unit convention.
-
-    Returns:
-        ``(lat, lon, sample_dims)`` where *lat* and *lon* are flattened
-        float64 arrays and *sample_dims* are the dimensions of the
-        coordinate variables (e.g. ``("along_track", "across_track")``).
-
-    Raises:
-        ValueError: When latitude and longitude disagree in dims/shape.
-    """
-    lat_var = ds[_find_coord_name(ds, _LAT_NAMES, lat_name, "latitude")]
-    lon_var = ds[_find_coord_name(ds, _LON_NAMES, lon_name, "longitude")]
-
-    if tuple(lat_var.dims) != tuple(lon_var.dims):
-        raise ValueError(
-            "Latitude and longitude must share the same dimensions, got "
-            f"{tuple(lat_var.dims)} vs {tuple(lon_var.dims)}."
-        )
-    if lat_var.shape != lon_var.shape:
-        raise ValueError(
-            "Latitude and longitude must share the same shape, got "
-            f"{lat_var.shape} vs {lon_var.shape}."
-        )
-
-    lat = _normalise_angle_units(
-        np.asarray(lat_var.values, dtype=np.float64).ravel(), source_units
-    )
-    lon = _canonical_lon(
-        _normalise_angle_units(
-            np.asarray(lon_var.values, dtype=np.float64).ravel(), source_units
-        )
-    )
-    return lat, lon, tuple(map(str, lat_var.dims))
-
-
-# ===================================================================
-# Fill-value handling
-# ===================================================================
-
-
-def _declared_fill_value(da: xr.DataArray) -> float | None:
-    """Return the fill value declared on *da*, if any.
-
-    Both ``attrs`` and ``encoding`` are inspected for ``_FillValue`` and
-    ``missing_value``.  No dtype-based guessing is performed: a fill
-    value is only honoured when the producer declared one.
-
-    Args:
-        da: Data variable.
-
-    Returns:
-        The declared fill value, or *None*.
-    """
-    for source in (da.attrs, da.encoding):
-        for attr in _FILL_ATTR_NAMES:
-            value = source.get(attr)
-            if value is not None:
-                return float(np.asarray(value).ravel()[0])
-    return None
-
-
-def _masked_float64(
-    da: xr.DataArray,
-    *,
-    fill_value: float | None,
-) -> FloatArray:
-    """Return *da* as float64 with invalid samples set to NaN.
-
-    Invalid samples are those equal to the fill value (explicit
-    *fill_value* argument, falling back to a declared ``_FillValue`` /
-    ``missing_value``) and, for floating-point input, non-finite values.
-
-    Args:
-        da: Data variable.
-        fill_value: Explicit fill value overriding declared attributes.
-            Note that ``0`` is a perfectly valid fill value here — the
-            check is ``is None``, never truthiness.
-
-    Returns:
-        Float64 array of the same shape with NaN at invalid samples.
-    """
-    fill = fill_value if fill_value is not None else _declared_fill_value(da)
-    raw = np.asarray(da.values)
-    values = raw.astype(np.float64)
-    invalid = ~np.isfinite(values)
-    if fill is not None and np.isfinite(fill):
-        # Compare in float64 so integer fills (255, -32768, 0, ...) and
-        # large CF float fills (9.9692e36) are both handled without
-        # dtype-cast surprises.
-        invalid |= values == float(fill)
-    values[invalid] = np.nan
-    return values
-
-
-# ===================================================================
-# Per-cell reductions
-# ===================================================================
-
-
-def _bin_simple(
-    group_idx: Int64Array,
-    values: FloatArray,
-    *,
-    n_cells: int,
-    func: str,
-) -> FloatArray:
-    """Apply a NaN-aware numpy-groupies reduction along the last axis."""
-    result = npg.aggregate(
-        group_idx,
-        values,
-        axis=-1,
-        func=func,
-        size=n_cells,
-        fill_value=np.nan,
-    )
-    return cast(FloatArray, np.asarray(result, dtype=np.float64))
-
-
-def _bin_mode(
-    group_idx: Int64Array,
-    values: FloatArray,
-    *,
-    n_cells: int,
-    max_classes: int = 256,
-) -> FloatArray:
-    """Per-cell mode for categorical data.
-
-    Ties are broken deterministically by the *lowest class value* (the
-    class loop runs over sorted unique values and a later class only wins
-    with a strictly greater count).
-
-    Args:
-        group_idx: Compact cell index per sample (last axis).
-        values: Sample values with NaN marking invalid samples.
-        n_cells: Number of cells.
-        max_classes: Guard against accidentally binning continuous data.
-
-    Returns:
-        Mode per cell; NaN where no valid sample exists.
-
-    Raises:
-        ValueError: When the number of distinct classes exceeds
-            *max_classes*.
-    """
-    classes = np.unique(values[np.isfinite(values)])
-    batch_shape = values.shape[:-1]
-    best_value = np.full((*batch_shape, n_cells), np.nan)
-    if classes.size == 0:
-        return best_value
-    if classes.size > max_classes:
-        raise ValueError(
-            f"Found {classes.size} distinct classes (> {max_classes}). "
-            "Mode binning is meant for categorical data; use agg='mean' "
-            "for continuous fields."
-        )
-    best_count = np.zeros((*batch_shape, n_cells), dtype=np.int64)
-    for cls in classes:
-        count = npg.aggregate(
-            group_idx,
-            (values == cls).astype(np.int64),
-            axis=-1,
-            func="sum",
-            size=n_cells,
-            fill_value=0,
-        ).astype(np.int64)
-        wins = count > best_count
-        best_value[wins] = cls
-        best_count[wins] = count[wins]
-    return best_value
-
-
-def _bin_count(
-    group_idx: Int64Array,
-    values: FloatArray,
-    *,
-    n_cells: int,
-) -> FloatArray:
-    """Count the number of valid samples per cell."""
-    result = npg.aggregate(
-        group_idx,
-        # numpy-groupies treats 1-D boolean input to ``sum`` as a
-        # logical ``any`` -- cast explicitly to count samples.
-        np.isfinite(values).astype(np.int64),
-        axis=-1,
-        func="sum",
-        size=n_cells,
-        fill_value=0,
-    )
-    return cast(FloatArray, np.asarray(result, dtype=np.float64))
-
 
 # ===================================================================
 # Public API
@@ -438,14 +195,16 @@ def bin_to_healpix(
         pyramid[lvl] = gd.coarsen_healpix(pyramid[lvl + 1], lvl)
     ```
     """
-    if not 0 <= level <= _MAX_NESTED_LEVEL:
-        raise ValueError(f"level must be within [0, {_MAX_NESTED_LEVEL}].")
+    if not 0 <= level <= MAX_NESTED_LEVEL:
+        raise ValueError(f"level must be within [0, {MAX_NESTED_LEVEL}].")
     if min_count < 1:
         raise ValueError("min_count must be at least 1.")
 
-    lat, lon, sample_dims = _resolve_point_coords(
+    lat, lon, sample_dims = resolve_point_coords(
         ds, lat_name=lat_name, lon_name=lon_name, source_units=source_units
     )
+    # Fail-fast: validate every aggregation before any data is loaded.
+    methods = resolve_methods(ds, sample_dims, agg)
 
     valid_coords = np.isfinite(lat) & np.isfinite(lon)
     if not valid_coords.any():
@@ -480,37 +239,32 @@ def bin_to_healpix(
             )
             continue
 
-        method = str(agg[var_name] if isinstance(agg, Mapping) else agg)
-        if method not in _AGG_TO_METHOD:
-            raise ValueError(
-                f"Unknown aggregation {method!r} for {var_name!r}. "
-                f"Supported: {sorted(_AGG_TO_METHOD)}."
-            )
+        method = methods[var_name]
 
         batch_dims = tuple(d for d in map(str, da.dims) if d not in sample_dims)
         arranged = da.transpose(*batch_dims, *sample_dims)
         fill = (fill_values or {}).get(var_name)
-        values = _masked_float64(arranged, fill_value=fill)
+        values = masked_float64(arranged, fill_value=fill)
         values = values.reshape(*arranged.shape[: len(batch_dims)], -1)
         values = values[..., valid_coords]
 
-        valid_count = _bin_count(group_idx, values, n_cells=n_cells)
+        valid_count = bin_count(group_idx, values, n_cells=n_cells)
 
         if method == "count":
             result = valid_count
         elif method == "mode":
-            result = _bin_mode(group_idx, values, n_cells=n_cells)
+            result = bin_mode(group_idx, values, n_cells=n_cells)
             result[valid_count < min_count] = np.nan
         else:
-            result = _bin_simple(
-                group_idx, values, n_cells=n_cells, func=_NPG_FUNC[method]
+            result = bin_simple(
+                group_idx, values, n_cells=n_cells, func=NPG_FUNC[method]
             )
             result[valid_count < min_count] = np.nan
 
         attrs = {
-            key: value for key, value in da.attrs.items() if key not in _FILL_ATTR_NAMES
+            key: value for key, value in da.attrs.items() if key not in FILL_ATTR_NAMES
         }
-        attrs["grid_doctor_method"] = _AGG_TO_METHOD[method]
+        attrs["grid_doctor_method"] = AGG_TO_METHOD[method]
         binned[var_name] = xr.DataArray(result, dims=(*batch_dims, "cell"), attrs=attrs)
         if with_counts and method != "count":
             counts[f"{var_name}_count"] = xr.DataArray(
@@ -642,7 +396,7 @@ def _attach_sparse_coords(
     module, kwargs = _require_healpix_geo_module(nest)
     lon_deg, lat_deg = module.healpix_to_lonlat(cell_ids, level, **kwargs)
 
-    from .remap import _make_crs_variable
+    from ..remap import _make_crs_variable
 
     result = ds.assign_coords(
         cell=cell_ids,
